Add build instructions

+++ scripts/install-lua.sh
@@ -8,10 +8,12 @@ LUA_ARCHIVE=tmp/lua-$LUA_VERSION.tar.gz
 cd "$(dirname "$0")/.."
 
 mkdir -p tmp
-
-rm -rf $LUA_ARCHIVE
-curl --retry 2 --connect-timeout 30 --location https://github.com/lua/lua/archive/refs/tags/v$LUA_VERSION.tar.gz -o $LUA_ARCHIVE
 rm -rf lua/src
+
+if [ ! -f $LUA_ARCHIVE ]; then
+	curl --retry 2 --connect-timeout 30 --location https://github.com/lua/lua/archive/refs/tags/v$LUA_VERSION.tar.gz -o $LUA_ARCHIVE
+fi
+
 mkdir -p lua/src
 tar xf $LUA_ARCHIVE --strip-components=1 -C lua/src
 cd lua
